rename HeroMenu in SocialMenu
