+++ components/About/Teaser/index.js
